Simplify topics return
