--- client/src/containers/widgets/financial/idb-operations/columns.tsx
@@ -1,48 +1,0 @@
-"use client";
-
-import { ColumnDef } from "@tanstack/react-table";
-import { LuPin } from "react-icons/lu";
-
-import { formatCurrency } from "@/lib/formats";
-
-import { IDBOperation } from "@/containers/widgets/financial/types";
-
-export const columns: ColumnDef<IDBOperation>[] = [
-  {
-    accessorKey: "opername",
-    header: "Name",
-    sortingFn: "alphanumericCaseSensitive",
-    minSize: 250,
-    cell(props) {
-      const v = props.getValue<string>();
-      return (
-        <div className="flex space-x-2">
-          <LuPin className="mt-px shrink-0 text-cyan-600" />
-          <span>{v}</span>
-        </div>
-      );
-    },
-  },
-  {
-    accessorKey: "opertypenm",
-    header: "Type",
-    sortingFn: "alphanumericCaseSensitive",
-    maxSize: 100,
-  },
-  {
-    accessorKey: "sector",
-    header: "Sector",
-    sortingFn: "alphanumericCaseSensitive",
-    maxSize: 120,
-  },
-  {
-    accessorKey: "idbamount",
-    header: "Funding",
-    sortingFn: "alphanumericCaseSensitive",
-    maxSize: 90,
-    cell(props) {
-      const v = props.getValue<number>();
-      return <span>{formatCurrency(v)}</span>;
-    },
-  },
-];
