--- client/src/containers/widgets/overview/altitude.tsx
@@ -1,86 +1,0 @@
-"use client";
-
-import { treemapDice } from "@visx/hierarchy";
-import { HierarchyNode, HierarchyRectangularNode } from "@visx/hierarchy/lib/types";
-
-import { formatPercentage } from "@/lib/formats";
-import { useLocationGeometry } from "@/lib/location";
-import { useGetRasterAnalysis } from "@/lib/query";
-
-import { useSyncLocation } from "@/app/store";
-
-import { ELEVATION_RANGES, ElevationRangeIds } from "@/constants/colors";
-
-import { Card, CardTitle, CardLoader } from "@/containers/card";
-
-import MarimekkoChart, { Data } from "@/components/charts/marimekko";
-
-export default function WidgetAltitude() {
-  const [location] = useSyncLocation();
-
-  const GEOMETRY = useLocationGeometry(location, {
-    wkid: 4326,
-  });
-
-  const query = useGetRasterAnalysis(
-    {
-      id: "elevation_ranges",
-      polygon: GEOMETRY,
-      statistics: ["frac", "unique"],
-    },
-    {
-      enabled: !!GEOMETRY,
-
-      select(data): Data[] {
-        const values = data.features.map((f) => {
-          if (f.properties.unique && f.properties.frac) {
-            const { frac, unique } = f.properties;
-
-            const us = unique.map((u, index) => {
-              const e = ELEVATION_RANGES[`${u}` as ElevationRangeIds];
-              return {
-                id: e.label,
-                key: u,
-                parent: "root",
-                size: frac[index],
-                label: e.label,
-                color: e.color,
-              };
-            }, {});
-
-            return us.filter((u) => u.size > 0.001);
-          }
-
-          return [];
-        });
-
-        return values
-          .flat()
-          .toSorted((a, b) => a.key - b.key)
-          .map((d, i, arr) => {
-            return {
-              ...d,
-              percentage: (i + 1) / arr.length,
-            };
-          });
-      },
-    },
-  );
-
-  const FORMAT = (node: HierarchyRectangularNode<HierarchyNode<Data>>) => {
-    return formatPercentage(node?.value || 0, {
-      maximumFractionDigits: 0,
-    });
-  };
-
-  return (
-    <Card>
-      <CardTitle>Altitude</CardTitle>
-      <CardLoader query={[query]} className="h-12">
-        {!!query?.data && (
-          <MarimekkoChart data={query?.data} format={FORMAT} className="h-12" tile={treemapDice} />
-        )}
-      </CardLoader>
-    </Card>
-  );
-}
