--- client/src/containers/widgets/protection/indigenous-lands/types.ts
@@ -1,7 +1,0 @@
-export type IndigenousLand = {
-  FID: string;
-  categoria: string;
-  etnias: string;
-  Shape__Area: number;
-  Shape__Length: number;
-};
