--- client/src/containers/widgets/protection/protected-areas/types.ts
@@ -1,9 +1,0 @@
-export type ProtectedAreas = {
-  FID: string;
-  NAME: string;
-  DESIG: string;
-  DESIG_ENG: string;
-  IUCN_CAT: string;
-  Shape__Area: number;
-  Shape__Length: number;
-};
